META-373: Fixed compilation warning
<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/server/nx_vms_server/src/nx/vms/server/analytics/db/object_track_searcher.cpp
@@ -485,7 +485,7 @@ std::vector<ObjectTrack> ObjectTrackSearcher::loadTracks(
 
         tracks.push_back(std::move(*track));
 
-        if (limit > 0 && tracks.size() >= limit)
+        if (limit > 0 && (int) tracks.size() >= limit)
             break;
     }
